emotitask-backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
supabase>=2.15.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.26.0
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
openai>=1.3.0 